+++ engine/runtime/src/main/java/org/enso/interpreter/runtime/callable/UnresolvedSymbol.java
@@ -15,7 +15,7 @@ import org.enso.interpreter.runtime.state.data.EmptyMap;
 
 /** Simple runtime value representing a yet-unresolved by-name symbol. */
 @ExportLibrary(InteropLibrary.class)
-public class UnresolvedSymbol implements TruffleObject {
+public final class UnresolvedSymbol implements TruffleObject {
   private final String name;
   private final ModuleScope scope;